# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Musicfy API",
...
    "contact": {
      "name": "Musicfy Backend"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local"
    }
  ],
  "paths": {
...
    "/v1/uploads/audio": {
      "post": {
        "tags": [
          "Загрузка"
        ],
        "summary": "Загрузить аудио (для cover)",
        "operationId": "upload_audio_v1_uploads_audio_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_v1_uploads_audio_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/uploads/voice": {
      "post": {
        "tags": [
          "Загрузка"
        ],
        "summary": "Загрузить образец голоса",
        "operationId": "upload_voice_v1_uploads_voice_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_voice_v1_uploads_voice_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/uploads/source-video": {
      "post": {
        "tags": [
          "Загрузка"
        ],
        "summary": "Загрузить видео (для video)",
        "operationId": "upload_source_video_v1_uploads_source_video_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_source_video_v1_uploads_source_video_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/voices/consent": {
      "post": {
        "tags": [
          "Голоса"
        ],
        "summary": "Записать согласие на голос",
        "operationId": "create_consent_v1_voices_consent_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/voices": {
      "get": {
        "tags": [
          "Голоса"
        ],
        "summary": "Библиотека голосов",
        "operationId": "list_voices_v1_voices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VoiceProfileResponse"
                  },
                  "type": "array",
                  "title": "Response List Voices V1 Voices Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Голоса"
        ],
        "summary": "Клонировать голос",
        "operationId": "create_voice_v1_voices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVoiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VoiceProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/v1/videos": {
      "post": {
        "tags": [
          "Видео"
        ],
        "summary": "Создать AI music video",
        "operationId": "create_video_v1_videos_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 128
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateVideoRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/videos/{job_id}": {
      "get": {
        "tags": [
          "Видео"
        ],
        "summary": "Результат видео",
        "operationId": "get_video_v1_videos__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoResultResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetResponse": {
        "properties": {
          "assetId": {
            "type": "string",
            "title": "Assetid",
            "description": "ID ассета."
          },
          "url": {
            "type": "string",
            "title": "Url",
            "description": "URL для использования в запросах генерации."
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Тип: audio / video / voice_sample / source_video."
          },
          "mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mime"
          }
        },
        "type": "object",
        "required": [
          "assetId",
          "url",
          "kind"
        ],
        "title": "AssetResponse",
        "description": "Загруженный медиа-ассет."
      },
...
      "Body_upload_audio_v1_uploads_audio_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_audio_v1_uploads_audio_post"
      },
      "Body_upload_source_video_v1_uploads_source_video_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_source_video_v1_uploads_source_video_post"
      },
      "Body_upload_voice_v1_uploads_voice_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_voice_v1_uploads_voice_post"
      },
...
      "ConsentRequest": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "own_voice | third_party_authorized"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted",
            "description": "Подтверждение согласия (должно быть true)."
          },
          "statement": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Statement"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "accepted"
        ],
        "title": "ConsentRequest",
        "description": "Согласие на использование голоса. `accepted` обязан быть true."
      },
      "ConsentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "accepted": {
            "type": "boolean",
            "title": "Accepted"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "accepted"
        ],
        "title": "ConsentResponse"
      },
...
      "CreateVideoRequest": {
        "properties": {
          "audioUrl": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Audiourl",
            "description": "URL аудио (трека)."
          },
          "sourceVideoUrl": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Sourcevideourl",
            "description": "URL исходного видео (из /v1/uploads/source-video)."
          },
          "mode": {
            "type": "string",
            "maxLength": 32,
            "title": "Mode",
            "description": "Режим. В V1 — avatar_performance.",
            "default": "avatar_performance"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "required": [
          "audioUrl",
          "sourceVideoUrl"
        ],
        "title": "CreateVideoRequest",
        "description": "AI music video. `audioUrl` — готовый трек, `sourceVideoUrl` — аватар/исходное видео.",
        "example": {
          "audioUrl": "https://fal.media/files/.../song.mp3",
          "mode": "avatar_performance",
          "sourceVideoUrl": "https://fal.media/files/.../avatar.mp4"
        }
      },
      "CreateVoiceRequest": {
        "properties": {
          "sampleAssetUrl": {
            "type": "string",
            "maxLength": 1024,
            "minLength": 1,
            "title": "Sampleasseturl",
            "description": "URL образца голоса (из /v1/uploads/voice)."
          },
          "consentId": {
            "type": "string",
            "title": "Consentid",
            "description": "ID согласия из /v1/voices/consent."
          },
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "Имя профиля голоса."
          }
        },
        "type": "object",
        "required": [
          "sampleAssetUrl",
          "consentId"
        ],
        "title": "CreateVoiceRequest",
        "description": "Клонирование голоса. Сначала `POST /v1/voices/consent`, затем загрузка образца."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobAcceptedResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid",
            "description": "ID задачи для опроса статуса."
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Начальный статус (queued)."
          },
          "deduplicated": {
            "type": "boolean",
            "title": "Deduplicated",
            "description": "true — задача уже существовала (по Idempotency-Key).",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status"
        ],
        "title": "JobAcceptedResponse",
        "description": "Ответ на создание асинхронной задачи. Опрашивайте `GET /v1/jobs/{jobId}`."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VideoResultResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Статус задачи (см. GET /v1/jobs/{jobId})."
          },
          "videoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Videourl",
            "description": "URL готового видео (если completed)."
          },
          "createdAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "status"
        ],
        "title": "VideoResultResponse"
      },
      "VoiceProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "providerVoiceId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Providervoiceid"
          },
          "jobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jobid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "status",
          "createdAt"
        ],
        "title": "VoiceProfileResponse"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Админ-ключ (ADMIN_API_KEY)",
        "scheme": "bearer"
      }
    }
  },
...
}